[Enhancement #308] Extract common snapshots table code into one component.
